dependency cycle fixed
